move config to core lib
