+++ apps/domain-clients-example/src/lib/raydium/swap.ts
@@ -52,11 +52,6 @@ export const createClmmSwapInInstructions = async ({
     connection,
   });
 
-  const clmm = new Clmm({
-    scope: raydium,
-    moduleName: 'clmm',
-  });
-
   const { poolInfo, poolKeys, computePoolInfo } = await getPoolInfo(
     poolId,
     connection
