+++ packages/domain-clients-core/src/solana/config.ts
@@ -1,6 +1,32 @@
-import type { SolanaCluster, SolanaClusterId } from '@/solana';
+import {
+  type SolanaConfig,
+  type SolanaCluster,
+  isSolanaClusterId,
+} from '@/solana';
 
-export interface SolanaConfig {
+export const createSolanaDomainClientsConfig = ({
+  clusters,
+  defaultClusterId,
+}: {
   clusters: SolanaCluster[];
-  defaultClusterId: SolanaClusterId;
-}
+  defaultClusterId: string;
+}): SolanaConfig => {
+  if (!isSolanaClusterId(defaultClusterId)) {
+    throw new Error('Default cluster id must start with "solana:"');
+  }
+
+  // convert to format for wallet-ui (side-stepping 'moniker' feature)
+  const convertedClusters = clusters.map(cluster => {
+    return {
+      id: cluster.id,
+      label: cluster.label,
+      cluster: cluster.cluster,
+      urlOrMoniker: cluster.url,
+    };
+  });
+
+  return {
+    clusters: convertedClusters,
+    defaultClusterId: defaultClusterId,
+  };
+};
